requirements.txt
protobuf==3.20.0
streamlit
google-cloud
google-cloud-language
